chore: Refactor to stubber.publish.stubpackage.
