+++ tests/publish/test_multi_build.py
@@ -22,7 +22,7 @@ def test_build_no_change(mocker: MockerFixture, tmp_path: Path, pytestconfig: py
     # use the test config
     config = FakeConfig(tmp_path=tmp_path, rootpath=pytestconfig.rootpath)
     mocker.patch("stubber.publish.publish.CONFIG", config)
-    mocker.patch("stubber.publish.stubpacker.CONFIG", config)
+    mocker.patch("stubber.publish.stubpackage.CONFIG", config)
     m_is_changed: MagicMock = mocker.patch("stubber.publish.package.StubPackage.is_changed", autospec=True, return_value=False)  # type: ignore
 
     m_check: MagicMock = mocker.patch("stubber.publish.package.StubPackage.check", autospec=True, return_value=True)  # type: ignore
@@ -53,7 +53,7 @@ def test_build_changed(mocker: MockerFixture, tmp_path: Path, pytestconfig: pyte
     # use the test config
     config = FakeConfig(tmp_path=tmp_path, rootpath=pytestconfig.rootpath)
     mocker.patch("stubber.publish.publish.CONFIG", config)
-    mocker.patch("stubber.publish.stubpacker.CONFIG", config)
+    mocker.patch("stubber.publish.stubpackage.CONFIG", config)
     m_is_changed: MagicMock = mocker.patch(
         "stubber.publish.package.StubPackage.is_changed", autospec=True, return_value=False
     )
@@ -87,7 +87,7 @@ def test_build_force(mocker: MockerFixture, tmp_path: Path, pytestconfig: pytest
     # use the test config
     config = FakeConfig(tmp_path=tmp_path, rootpath=pytestconfig.rootpath)
     mocker.patch("stubber.publish.publish.CONFIG", config)
-    mocker.patch("stubber.publish.stubpacker.CONFIG", config)
+    mocker.patch("stubber.publish.stubpackage.CONFIG", config)
     m_is_changed: MagicMock = mocker.patch(
         "stubber.publish.package.StubPackage.is_changed", autospec=True, return_value=False
     )
